Rename command: clicontrib -> cliutils
